+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   ([PR #92](https://github.com/LBHackney-IT/LBH-frontend/pull/92))
 - Remove leaflet as a dependency that gets packaged up with lbh-frontend and instead require it in the app so that it can still be used but doesn't add weight to the package. Requires users to install leaflet as dependency
   ([PR #93](https://github.com/LBHackney-IT/LBH-frontend/pull/93))
+- Add panel component
+  ([PR #94](https://github.com/LBHackney-IT/LBH-frontend/pull/94))
 
 ## 1.8.3